patches,git-suse: Include "(partial)" commits
ATM these two tools will not include in their output the commit ids of
patches which are tagged with "Git-commit: [...] (partial)".

This has two consequences:
1)
It later leads git-fixes to consider as "missing" some commits which are
already partially applied. In my review those were all false positives.
They are removed with this patch.
2)
On the flipside, some partial backports which were not checked for followup
fixes previously now are. Therefore some false negatives are also solved.

Signed-off-by: Benjamin Poirier <bpoirier@suse.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>